Code refactory better efficiency and readability
Did some changes on how it works, speacilly by making the main variables
global, and setting specific functions for specific tasks, we still
have our "game" function, which is the main and runs a while loop, and
the "get_args", but included a "show" which was a merge of the block of
code that checked the user platform and clean screen with the
right command, and the block that showed remaining attempts, the found
and to be found letters. And a "gen_word" that generates the word.